Refactor main and evmProcessor
