+++ envs/.env.sepolia.example
@@ -7,10 +7,9 @@ DB_HOST=localhost
 PROCESSOR_PROMETHEUS_PORT=3000
 GQL_PORT=4350
 
-STATIC_FEE_DATA=0x00
-
 SHARED_CONFIG_URL=
 
+DOMAIN_TYPE=evm
 DOMAIN_ID=2
 DOMAIN_GATEWAY=
 DOMAIN_BRIDGE_ADDRESS=
