+++ src/main.ts
@@ -2,38 +2,19 @@
 The Licensed Work is (c) 2024 Sygma
 SPDX-License-Identifier: LGPL-3.0-only
 */
-import { TypeormDatabase } from "@subsquid/typeorm-store";
-import { ethers } from "ethers";
-import * as bridge from "./abi/bridge";
-import { processor } from "./evmProcessor";
+import { startEvmProcessing } from "./evmProcessor";
 import {
-  parseDeposit,
-  parseFailedHandlerExecution,
-  parseProposalExecution,
-} from "./evmIndexer/utils";
-import {
-  DecodedDepositLog,
-  DecodedFailedHandlerExecution,
-  DecodedProposalExecutionLog,
-} from "./evmIndexer/evmTypes";
-import {
-  getSharedConfig,
-  getSsmDomainConfig,
   getDomainConfig,
-  validateConfig,
+  getProcessorConfig,
+  DomainTypes,
+  getSharedConfig,
 } from "./config";
-import {
-  processDeposits,
-  processExecutions,
-  processFailedExecutions,
-} from "./evmIndexer/evmIndexer";
 import { logger } from "./utils/logger";
 
-async function startEVMProcessing(): Promise<void> {
+async function startProcessing(): Promise<void> {
+  const processorConfig = getProcessorConfig();
   const domainConfig = getDomainConfig();
-  validateConfig(domainConfig);
 
-  const provider = new ethers.JsonRpcProvider(domainConfig.rpcURL);
   const sharedConfig = await getSharedConfig(domainConfig.sharedConfigURL);
   const thisDomain = sharedConfig.domains.find(
     (domain) => domain.id == domainConfig.domainID
@@ -44,57 +25,21 @@ async function startEVMProcessing(): Promise<void> {
     );
   }
 
-  const substrateRpcUrlConfig = await getSsmDomainConfig(
-    domainConfig.supportedSubstrateRPCs
-  );
-
-  logger.info("Process initialization completed successfully.");
-
-  processor.run(
-    new TypeormDatabase({
-      stateSchema: process.env.DOMAIN_ID,
-      isolationLevel: "READ COMMITTED",
-    }),
-    async (ctx) => {
-      const deposits: DecodedDepositLog[] = [];
-      const executions: DecodedProposalExecutionLog[] = [];
-      const failedHandlerExecutions: DecodedFailedHandlerExecution[] = [];
-      for (const block of ctx.blocks) {
-        for (const log of block.logs) {
-          if (log.topics[0] === bridge.events.Deposit.topic) {
-            const event = bridge.events.Deposit.decode(log);
-            const toDomain = sharedConfig.domains.find(
-              (domain) => domain.id == event.destinationDomainID
-            );
-            deposits.push(
-              await parseDeposit(
-                log,
-                thisDomain,
-                toDomain!,
-                provider,
-                substrateRpcUrlConfig
-              )
-            );
-          } else if (log.topics[0] === bridge.events.ProposalExecution.topic) {
-            executions.push(parseProposalExecution(log, thisDomain));
-          } else if (
-            log.topics[0] === bridge.events.FailedHandlerExecution.topic
-          ) {
-            failedHandlerExecutions.push(
-              parseFailedHandlerExecution(log, thisDomain)
-            );
-          }
-        }
-      }
-
-      await processDeposits(ctx, deposits);
-      await processExecutions(ctx, executions);
-      await processFailedExecutions(ctx, failedHandlerExecutions);
-    }
-  );
+  switch (domainConfig.domainType) {
+    case DomainTypes.EVM:
+      await startEvmProcessing(
+        processorConfig,
+        domainConfig,
+        sharedConfig,
+        thisDomain
+      );
+      break;
+    default:
+      throw new Error("Unsupported domain type");
+  }
 }
 
-startEVMProcessing()
+startProcessing()
   .then(() => {
     logger.info("Processing started successfully.");
   })
